Firestorm: Prepare for i2c test
