+++ src/platform/storm/lib.rs
@@ -14,19 +14,16 @@ use hil::spi_master::SpiMaster;
 use drivers::timer::AlarmToTimer;
 use drivers::virtual_alarm::{MuxAlarm, VirtualMuxAlarm};
 
+#[macro_use]
+pub mod io;
+
 // HAL unit tests. To enable a particular unit test, uncomment
 // the module here and uncomment the call to start the test in
 // the init function below.
 //mod gpio_dummy;
 //mod spi_dummy;
-//mod spi_driver;
+//mod i2c_dummy;
 
-
-#[allow(unused_variables,dead_code)]
-pub struct DummyCB {
-    val: u8
-}
- 
 static mut spi_read_buf:  [u8; 64] = [0; 64];
 static mut spi_write_buf: [u8; 64] = [0; 64];
 
@@ -122,6 +119,10 @@ pub unsafe fn init<'a>() -> &'a mut Firestorm {
     static_init!(spi: drivers::spi::Spi<'static, sam4l::spi::Spi> =
                       drivers::spi::Spi::new(&mut sam4l::spi::SPI));
     spi.config_buffers(&mut spi_read_buf, &mut spi_write_buf);
+    sam4l::spi::SPI.set_active_peripheral(sam4l::spi::Peripheral::Peripheral1);
+    sam4l::spi::SPI.init(spi as &hil::spi_master::SpiCallback);
+    sam4l::spi::SPI.enable();
+
 
     static_init!(firestorm : Firestorm = Firestorm {
         chip: sam4l::chip::Sam4l::new(),
@@ -162,67 +163,13 @@ pub unsafe fn init<'a>() -> &'a mut Firestorm {
     // Firestorm's pin 8 changes value (e.g., connect a push button to pin 8 and
     // press toggle it).
     //gpio_dummy::gpio_dummy_test();
+    
+    // Uncommenting the following line will test the I2C
+    //i2c_dummy::i2c_dummy_test(&mut sam4l::i2c::I2C2);
 
-    sam4l::spi::SPI.set_active_peripheral(sam4l::spi::Peripheral::Peripheral1);
-    sam4l::spi::SPI.init(spi as &hil::spi_master::SpiCallback);
-    sam4l::spi::SPI.enable();
 
     firestorm.console.initialize();
+
     firestorm
-}
-
-use core::fmt::Arguments;
-
-#[cfg(not(test))]
-#[lang="panic_fmt"]
-#[no_mangle]
-pub unsafe extern fn rust_begin_unwind(args: Arguments,
-    file: &'static str, line: u32) -> ! {
-
-    static mut NUM : usize = 0;
-
-    use hil::uart::UART;
-    use core::fmt::*;
-    use support::nop;
-
-    NUM += 1;
-
-    sam4l::usart::USART3.configure(sam4l::usart::USARTParams {
-        baud_rate: 115200,
-        data_bits: 8,
-        parity: hil::uart::Parity::None
-    });
-    sam4l::usart::USART3.enable_tx();
-
-    struct Writer;
-
-    impl Write for Writer {
-        fn write_str(&mut self, s: &str) -> ::core::fmt::Result {
-            unsafe {
-                for c in s.bytes() {
-                    sam4l::usart::USART3.send_byte(c);
-                }
-            }
-            Ok(())
-        }
-    }
-
-    let mut writer = Writer;
-    let _ = writer.write_fmt(format_args!("Kernel panic at {}:{}:\r\n\t\"", file, line));
-    let _ = write(&mut writer, args);
-    let _ = writer.write_str("\"\r\n");
-
-    let led = &sam4l::gpio::PC[10];
-    led.enable_output();
-    loop {
-        for _ in 0..1000000 {
-            led.set();
-            nop();
-        }
-        for _ in 0..1000000 {
-            led.clear();
-            nop();
-        }
-    }
 }
